cleaning up props and usage
